+++ asf_search/Pair.py
@@ -1,7 +1,9 @@
 import importlib.util
 import math
+from typing import Union
 
 from .ASFProduct import ASFProduct
+from .S1MultiBurstProduct import S1MultiBurstProduct
 from .baseline import calculate_perpendicular_baselines
 from .exceptions import CoherenceEstimationError
 import pytz
@@ -27,8 +29,7 @@ except ImportError:
 
 class Pair:
     """
-    A Pair is comprised of a reference scene and a secondary scene. These scenes typically intersect geographically,
-    but that is not a requirement. When a pair is created, its perpendicular and temporal baselines are calculated
+    A Pair is comprised of a reference scene and a secondary scene. When a pair is created, its perpendicular and temporal baselines are calculated
     and stored in the self.perpendicular_baseline and self.temporal_baseline member variables.
 
     Two pairs are equivalent if they have matching reference and secondary dates
@@ -36,19 +37,30 @@ class Pair:
     ref: ASFProduct for reference scene
     sec: ASFProduct for secondary scene
     """
-    def __init__(self, ref: ASFProduct, sec: ASFProduct):
+    def __init__(self, 
+            ref: Union[ASFProduct, S1MultiBurstProduct], 
+            sec: Union[ASFProduct, S1MultiBurstProduct]):
+        
+
         self.ref = ref
-        self.sec = sec
-        self.id = (ref.properties['sceneName'], sec.properties['sceneName'])
+        self.sec = sec          
+        
+        self.id = (self.ref.properties['sceneName'], self.sec.properties['sceneName'])
+        if isinstance(ref, ASFProduct) and isinstance(sec, ASFProduct):
+            self.perpendicular_baseline = calculate_perpendicular_baselines(
+                self.ref.properties['sceneName'], 
+                [self.sec, self.ref])[0].properties['perpendicularBaseline']
+            
+        elif isinstance(ref, S1MultiBurstProduct) and isinstance(sec, S1MultiBurstProduct):
+            perpendicular_baselines = [calculate_perpendicular_baselines(ref.properties['sceneName'], [sec, ref])[0].properties['perpendicularBaseline'] 
+                                       for ref, sec in zip(self.ref.geo_reference_bursts, self.sec.geo_reference_bursts)]
+            # Use the largest burst perpendicular baseline to represent the S1MultiBurstProduct
+            self.perpendicular_baseline = max((x for x in perpendicular_baselines if x is not None), default=None)
 
-        self.perpendicular_baseline = calculate_perpendicular_baselines(
-            ref.properties['sceneName'], 
-            [sec, ref])[0].properties['perpendicularBaseline']
-
-        self.ref_time = parse_datetime(ref.properties["startTime"])
+        self.ref_time = parse_datetime(self.ref.properties["startTime"])
         if self.ref_time.tzinfo is None:
             self.ref_time = pytz.utc.localize(self.ref_time)
-        self.sec_time = parse_datetime(sec.properties["startTime"])
+        self.sec_time = parse_datetime(self.sec.properties["startTime"])
         if self.sec_time.tzinfo is None:
             self.sec_time = pytz.utc.localize(self.sec_time)