+++ asf_search/constants/__init__.py
@@ -11,3 +11,4 @@ from .INTERNAL import *  # noqa: F403 F401
 from .DATASET import *  # noqa: F403 F401
 from .RANGE_BANDWIDTH import *  # noqa: F403 F401
 from .PRODUCTION_CONFIGURATION import * # noqa: F403 F401
+from .MATURITIES import * # noqa: F403 F401